+++ supersuit/aec_vector/__init__.py
@@ -1,2 +1,2 @@
 from .base_aec_vec_env import VectorAECEnv
-from .create import vectorize_aec_env
+from .create import vectorize_aec_env_v0
