+++ supersuit/generic_wrappers/lambda_wrappers/__init__.py
@@ -1,3 +1,3 @@
-from .action_lambda import action_lambda_v1
+from .action_lambda import action_lambda_v1, action_lambda_v0
 from .observation_lambda import observation_lambda_v0
 from .reward_lambda import reward_lambda_v0
