+++ supersuit/vector/vector_constructors.py
@@ -1,6 +1,6 @@
 import gym
 import cloudpickle
-from .vector import MakeCPUAsyncConstructor, MarkovVectorEnv
+from . import MakeCPUAsyncConstructor, MarkovVectorEnv
 from pettingzoo.utils.env import AECEnv, ParallelEnv
 import warnings
 
@@ -19,14 +19,14 @@ def warn_not_gym_env(env, fn_name):
         warnings.warn(f"{fn_name} took in an environment which does not inherit from gym.Env. Note that gym_vec_env only takes in gym-style environments, not pettingzoo environments.")
 
 
-def gym_vec_env(env, num_envs, multiprocessing=False):
+def gym_vec_env_v0(env, num_envs, multiprocessing=False):
     warn_not_gym_env(env, "gym_vec_env")
     args = vec_env_args(env, num_envs)
     constructor = gym.vector.AsyncVectorEnv if multiprocessing else gym.vector.SyncVectorEnv
     return constructor(*args)
 
 
-def stable_baselines_vec_env(env, num_envs, multiprocessing=False):
+def stable_baselines_vec_env_v0(env, num_envs, multiprocessing=False):
     import stable_baselines
 
     warn_not_gym_env(env, "stable_baselines_vec_env")
@@ -35,7 +35,7 @@ def stable_baselines_vec_env(env, num_envs, multiprocessing=False):
     return constructor(*args)
 
 
-def stable_baselines3_vec_env(env, num_envs, multiprocessing=False):
+def stable_baselines3_vec_env_v0(env, num_envs, multiprocessing=False):
     import stable_baselines3
 
     warn_not_gym_env(env, "stable_baselines3_vec_env")
@@ -44,7 +44,7 @@ def stable_baselines3_vec_env(env, num_envs, multiprocessing=False):
     return constructor(*args)
 
 
-def concat_vec_envs(vec_env, num_vec_envs, num_cpus=0, base_class='gym'):
+def concat_vec_envs_v0(vec_env, num_vec_envs, num_cpus=0, base_class='gym'):
     num_cpus = min(num_cpus, num_vec_envs)
     vec_env = MakeCPUAsyncConstructor(num_cpus)(*vec_env_args(vec_env, num_vec_envs))
 
@@ -60,6 +60,6 @@ def concat_vec_envs(vec_env, num_vec_envs, num_cpus=0, base_class='gym'):
         raise ValueError("supersuit_vec_env only supports 'gym', 'stable_baselines', and 'stable_baselines3' for its base_class")
 
 
-def pettingzoo_env_to_vec_env(parallel_env):
+def pettingzoo_env_to_vec_env_v0(parallel_env):
     assert isinstance(parallel_env, ParallelEnv), "pettingzoo_env_to_vec_env takes in a pettingzoo ParallelEnv. Can create a parallel_env with pistonball.parallel_env() or convert it from an AEC env with `from pettingzoo.utils.conversions import to_parallel; to_parallel(env)``"
     return MarkovVectorEnv(parallel_env)
